% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{lambda}
\title{intensity object suitable for use in Cox simulation}
\usage{
lambda(x, kernel = "gauss", sigma = 0.5, type = "sum", alpha = 1)
}
\arguments{
\item{x}{matrix of coordinates}

\item{kernel}{type of kernel to use, "gauss" or "step"}

\item{sigma}{kernel parameter, sd of gauss or range of step}

\item{type}{"sum" or "product", see details}

\item{alpha}{additional scaling parameter}
}
\description{
intensity object suitable for use in Cox simulation
}
\details{
For type="sum", the field is 

v(u) = alpha * sum k(u-x)

where k is a kernel, and type="product"

v(u) = exp(alpha[1]) * prod [ 1 + alpha[2]*k(u-x)/k(0) ]

following Jalilian et al. 2015. The kernels are normalised densities in the dimension of x.

This function merely creates an field object. Use "coxintensity2matrix" to evaluate it at 
arbitrary locations, or "rcox" to simulate uniform/poisson process on it.
}
