% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{plot.coxintensity}
\alias{plot.coxintensity}
\title{plot lambda-object}
\usage{
\method{plot}{coxintensity}(x, bbox = cbind(0:1, 0:1), ..., nx = 100)
}
\arguments{
\item{bbox}{bounding box for the realisation, col-wise}

\item{...}{passed on to plot.im0()}
}
\description{
plot lambda-object
}
