% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-lambda.R
\name{print.im0}
\alias{print.im0}
\title{print for coxintensitymatrix}
\usage{
\method{print}{im0}(x, ...)
}
\arguments{
\item{x}{coxintensity2matrix result}

\item{...}{ignored}
}
\description{
print for coxintensitymatrix
}
