% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcox.R
\name{rcox}
\alias{rcox}
\title{Simulate Cox process in 2- or 3-dimensional box}
\usage{
rcox(lambda, n, bbox = cbind(c(0, 1), c(0, 1), c(0, 1)), iter = 10000,
  verb = FALSE)
}
\description{
Simulate the spatial Cox point process in 2- or 3-dimensional box 
(rectangular cuboid). Three algorithms are available:
}
